extract migrations
