+++ composeApp/src/androidMain/kotlin/it/fast4x/rimusic/Database.kt
@@ -1,25 +1,16 @@
 package it.fast4x.rimusic
 
-import android.content.ContentValues
-import android.content.Context
 import android.database.SQLException
-import android.database.sqlite.SQLiteDatabase.CONFLICT_IGNORE
 import android.os.Parcel
-import androidx.core.database.getFloatOrNull
-import androidx.lifecycle.ViewModelProvider.NewInstanceFactory.Companion.instance
 import androidx.media3.common.MediaItem
 import androidx.media3.common.util.UnstableApi
 import androidx.room.AutoMigration
 import androidx.room.Dao
 import androidx.room.Delete
-import androidx.room.DeleteColumn
-import androidx.room.DeleteTable
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.RawQuery
-import androidx.room.RenameColumn
-import androidx.room.RenameTable
 import androidx.room.RewriteQueriesToDropUnusedColumns
 import androidx.room.Room
 import androidx.room.RoomDatabase
@@ -29,10 +20,7 @@ import androidx.room.TypeConverter
 import androidx.room.TypeConverters
 import androidx.room.Update
 import androidx.room.Upsert
-import androidx.room.migration.AutoMigrationSpec
-import androidx.room.migration.Migration
 import androidx.sqlite.db.SimpleSQLiteQuery
-import androidx.sqlite.db.SupportSQLiteDatabase
 import androidx.sqlite.db.SupportSQLiteQuery
 import it.fast4x.rimusic.enums.AlbumSortBy
 import it.fast4x.rimusic.enums.ArtistSortBy
@@ -42,8 +30,8 @@ import it.fast4x.rimusic.enums.SongSortBy
 import it.fast4x.rimusic.enums.SortOrder
 import it.fast4x.rimusic.models.Album
 import it.fast4x.rimusic.models.Artist
-import it.fast4x.rimusic.models.SongWithContentLength
 import it.fast4x.rimusic.models.Event
+import it.fast4x.rimusic.models.EventWithSong
 import it.fast4x.rimusic.models.Format
 import it.fast4x.rimusic.models.Info
 import it.fast4x.rimusic.models.Lyrics
@@ -55,14 +43,22 @@ import it.fast4x.rimusic.models.SearchQuery
 import it.fast4x.rimusic.models.Song
 import it.fast4x.rimusic.models.SongAlbumMap
 import it.fast4x.rimusic.models.SongArtistMap
-import it.fast4x.rimusic.models.SongPlaylistMap
-import it.fast4x.rimusic.models.SortedSongPlaylistMap
-import it.fast4x.rimusic.models.EventWithSong
 import it.fast4x.rimusic.models.SongEntity
+import it.fast4x.rimusic.models.SongPlaylistMap
+import it.fast4x.rimusic.models.SongWithContentLength
+import it.fast4x.rimusic.models.SortedSongPlaylistMap
 import it.fast4x.rimusic.service.LOCAL_KEY_PREFIX
-import kotlin.jvm.Throws
 import kotlinx.coroutines.flow.Flow
 import me.knighthat.appContext
+import me.knighthat.database.migrator.From10To11Migration
+import me.knighthat.database.migrator.From11To12Migration
+import me.knighthat.database.migrator.From14To15Migration
+import me.knighthat.database.migrator.From20To21Migration
+import me.knighthat.database.migrator.From21To22Migration
+import me.knighthat.database.migrator.From22To23Migration
+import me.knighthat.database.migrator.From3To4Migration
+import me.knighthat.database.migrator.From7To8Migration
+import me.knighthat.database.migrator.From8To9Migration
 
 
 @Dao
@@ -1507,13 +1503,13 @@ interface Database {
     autoMigrations = [
         AutoMigration(from = 1, to = 2),
         AutoMigration(from = 2, to = 3),
-        AutoMigration(from = 3, to = 4, spec = DatabaseInitializer.From3To4Migration::class),
+        AutoMigration(from = 3, to = 4, spec = From3To4Migration::class),
         AutoMigration(from = 4, to = 5),
         AutoMigration(from = 5, to = 6),
         AutoMigration(from = 6, to = 7),
-        AutoMigration(from = 7, to = 8, spec = DatabaseInitializer.From7To8Migration::class),
+        AutoMigration(from = 7, to = 8, spec = From7To8Migration::class),
         AutoMigration(from = 9, to = 10),
-        AutoMigration(from = 11, to = 12, spec = DatabaseInitializer.From11To12Migration::class),
+        AutoMigration(from = 11, to = 12, spec = From11To12Migration::class),
         AutoMigration(from = 12, to = 13),
         AutoMigration(from = 13, to = 14),
         AutoMigration(from = 15, to = 16),
@@ -1521,8 +1517,8 @@ interface Database {
         AutoMigration(from = 17, to = 18),
         AutoMigration(from = 18, to = 19),
         AutoMigration(from = 19, to = 20),
-        AutoMigration(from = 20, to = 21, spec = DatabaseInitializer.From20To21Migration::class),
-        AutoMigration(from = 21, to = 22, spec = DatabaseInitializer.From21To22Migration::class),
+        AutoMigration(from = 20, to = 21, spec = From20To21Migration::class),
+        AutoMigration(from = 21, to = 22, spec = From21To22Migration::class),
     ],
 )
 @TypeConverters(Converters::class)
@@ -1551,148 +1547,6 @@ abstract class DatabaseInitializer protected constructor() : RoomDatabase() {
 
         fun reload() = synchronized(this) {
             Instance = getDatabase()
-        }
-    }
-
-    @DeleteTable.Entries(DeleteTable(tableName = "QueuedMediaItem"))
-    class From3To4Migration : AutoMigrationSpec
-
-    @RenameColumn.Entries(RenameColumn("Song", "albumInfoId", "albumId"))
-    class From7To8Migration : AutoMigrationSpec
-
-    class From8To9Migration : Migration(8, 9) {
-        override fun migrate(db: SupportSQLiteDatabase) {
-            db.query(SimpleSQLiteQuery("SELECT DISTINCT browseId, text, Info.id FROM Info JOIN Song ON Info.id = Song.albumId;"))
-                .use { cursor ->
-                    val albumValues = ContentValues(2)
-                    while (cursor.moveToNext()) {
-                        albumValues.put("id", cursor.getString(0))
-                        albumValues.put("title", cursor.getString(1))
-                        db.insert("Album", CONFLICT_IGNORE, albumValues)
-
-                        db.execSQL(
-                            "UPDATE Song SET albumId = '${cursor.getString(0)}' WHERE albumId = ${
-                                cursor.getLong(
-                                    2
-                                )
-                            }"
-                        )
-                    }
-                }
-
-            db.query(SimpleSQLiteQuery("SELECT GROUP_CONCAT(text, ''), SongWithAuthors.songId FROM Info JOIN SongWithAuthors ON Info.id = SongWithAuthors.authorInfoId GROUP BY songId;"))
-                .use { cursor ->
-                    val songValues = ContentValues(1)
-                    while (cursor.moveToNext()) {
-                        songValues.put("artistsText", cursor.getString(0))
-                        db.update(
-                            "Song",
-                            CONFLICT_IGNORE,
-                            songValues,
-                            "id = ?",
-                            arrayOf(cursor.getString(1))
-                        )
-                    }
-                }
-
-            db.query(SimpleSQLiteQuery("SELECT browseId, text, Info.id FROM Info JOIN SongWithAuthors ON Info.id = SongWithAuthors.authorInfoId WHERE browseId NOT NULL;"))
-                .use { cursor ->
-                    val artistValues = ContentValues(2)
-                    while (cursor.moveToNext()) {
-                        artistValues.put("id", cursor.getString(0))
-                        artistValues.put("name", cursor.getString(1))
-                        db.insert("Artist", CONFLICT_IGNORE, artistValues)
-
-                        db.execSQL(
-                            "UPDATE SongWithAuthors SET authorInfoId = '${cursor.getString(0)}' WHERE authorInfoId = ${
-                                cursor.getLong(
-                                    2
-                                )
-                            }"
-                        )
-                    }
-                }
-
-            db.execSQL("INSERT INTO SongArtistMap(songId, artistId) SELECT songId, authorInfoId FROM SongWithAuthors")
-
-            db.execSQL("DROP TABLE Info;")
-            db.execSQL("DROP TABLE SongWithAuthors;")
-        }
-    }
-
-    class From10To11Migration : Migration(10, 11) {
-        override fun migrate(db: SupportSQLiteDatabase) {
-            db.query(SimpleSQLiteQuery("SELECT id, albumId FROM Song;")).use { cursor ->
-                val songAlbumMapValues = ContentValues(2)
-                while (cursor.moveToNext()) {
-                    songAlbumMapValues.put("songId", cursor.getString(0))
-                    songAlbumMapValues.put("albumId", cursor.getString(1))
-                    db.insert("SongAlbumMap", CONFLICT_IGNORE, songAlbumMapValues)
-                }
-            }
-
-            db.execSQL("CREATE TABLE IF NOT EXISTS `Song_new` (`id` TEXT NOT NULL, `title` TEXT NOT NULL, `artistsText` TEXT, `durationText` TEXT NOT NULL, `thumbnailUrl` TEXT, `lyrics` TEXT, `likedAt` INTEGER, `totalPlayTimeMs` INTEGER NOT NULL, `loudnessDb` REAL, `contentLength` INTEGER, PRIMARY KEY(`id`))")
-
-            db.execSQL("INSERT INTO Song_new(id, title, artistsText, durationText, thumbnailUrl, lyrics, likedAt, totalPlayTimeMs, loudnessDb, contentLength) SELECT id, title, artistsText, durationText, thumbnailUrl, lyrics, likedAt, totalPlayTimeMs, loudnessDb, contentLength FROM Song;")
-            db.execSQL("DROP TABLE Song;")
-            db.execSQL("ALTER TABLE Song_new RENAME TO Song;")
-        }
-    }
-
-    @RenameTable("SongInPlaylist", "SongPlaylistMap")
-    @RenameTable("SortedSongInPlaylist", "SortedSongPlaylistMap")
-    class From11To12Migration : AutoMigrationSpec
-
-    class From14To15Migration : Migration(14, 15) {
-        override fun migrate(db: SupportSQLiteDatabase) {
-            db.query(SimpleSQLiteQuery("SELECT id, loudnessDb, contentLength FROM Song;"))
-                .use { cursor ->
-                    val formatValues = ContentValues(3)
-                    while (cursor.moveToNext()) {
-                        formatValues.put("songId", cursor.getString(0))
-                        formatValues.put("loudnessDb", cursor.getFloatOrNull(1))
-                        formatValues.put("contentLength", cursor.getFloatOrNull(2))
-                        db.insert("Format", CONFLICT_IGNORE, formatValues)
-                    }
-                }
-
-            db.execSQL("CREATE TABLE IF NOT EXISTS `Song_new` (`id` TEXT NOT NULL, `title` TEXT NOT NULL, `artistsText` TEXT, `durationText` TEXT NOT NULL, `thumbnailUrl` TEXT, `lyrics` TEXT, `likedAt` INTEGER, `totalPlayTimeMs` INTEGER NOT NULL, PRIMARY KEY(`id`))")
-
-            db.execSQL("INSERT INTO Song_new(id, title, artistsText, durationText, thumbnailUrl, lyrics, likedAt, totalPlayTimeMs) SELECT id, title, artistsText, durationText, thumbnailUrl, lyrics, likedAt, totalPlayTimeMs FROM Song;")
-            db.execSQL("DROP TABLE Song;")
-            db.execSQL("ALTER TABLE Song_new RENAME TO Song;")
-        }
-    }
-
-    @DeleteColumn.Entries(
-        DeleteColumn("Artist", "shuffleVideoId"),
-        DeleteColumn("Artist", "shufflePlaylistId"),
-        DeleteColumn("Artist", "radioVideoId"),
-        DeleteColumn("Artist", "radioPlaylistId"),
-    )
-    class From20To21Migration : AutoMigrationSpec
-
-    @DeleteColumn.Entries(DeleteColumn("Artist", "info"))
-    class From21To22Migration : AutoMigrationSpec
-
-    class From22To23Migration : Migration(22, 23) {
-        override fun migrate(db: SupportSQLiteDatabase) {
-            db.execSQL("CREATE TABLE IF NOT EXISTS Lyrics (`songId` TEXT NOT NULL, `fixed` TEXT, `synced` TEXT, PRIMARY KEY(`songId`), FOREIGN KEY(`songId`) REFERENCES `Song`(`id`) ON UPDATE NO ACTION ON DELETE CASCADE)")
-
-            db.query(SimpleSQLiteQuery("SELECT id, lyrics, synchronizedLyrics FROM Song;")).use { cursor ->
-                val lyricsValues = ContentValues(3)
-                while (cursor.moveToNext()) {
-                    lyricsValues.put("songId", cursor.getString(0))
-                    lyricsValues.put("fixed", cursor.getString(1))
-                    lyricsValues.put("synced", cursor.getString(2))
-                    db.insert("Lyrics", CONFLICT_IGNORE, lyricsValues)
-                }
-            }
-
-            db.execSQL("CREATE TABLE IF NOT EXISTS Song_new (`id` TEXT NOT NULL, `title` TEXT NOT NULL, `artistsText` TEXT, `durationText` TEXT, `thumbnailUrl` TEXT, `likedAt` INTEGER, `totalPlayTimeMs` INTEGER NOT NULL, PRIMARY KEY(`id`))")
-            db.execSQL("INSERT INTO Song_new(id, title, artistsText, durationText, thumbnailUrl, likedAt, totalPlayTimeMs) SELECT id, title, artistsText, durationText, thumbnailUrl, likedAt, totalPlayTimeMs FROM Song;")
-            db.execSQL("DROP TABLE Song;")
-            db.execSQL("ALTER TABLE Song_new RENAME TO Song;")
         }
     }
 }
