+++ composeApp/src/androidMain/kotlin/me/knighthat/database/migrator/From11To12Migration.kt
@@ -1,0 +1,8 @@
+package me.knighthat.database.migrator
+
+import androidx.room.RenameTable
+import androidx.room.migration.AutoMigrationSpec
+
+@RenameTable("SongInPlaylist", "SongPlaylistMap")
+@RenameTable("SortedSongInPlaylist", "SortedSongPlaylistMap")
+class From11To12Migration : AutoMigrationSpec
