+++ composeApp/src/androidMain/kotlin/me/knighthat/database/migrator/From20To21Migration.kt
@@ -1,0 +1,12 @@
+package me.knighthat.database.migrator
+
+import androidx.room.DeleteColumn
+import androidx.room.migration.AutoMigrationSpec
+
+@DeleteColumn.Entries(
+    DeleteColumn( "Artist", "shuffleVideoId" ),
+    DeleteColumn( "Artist", "shufflePlaylistId" ),
+    DeleteColumn( "Artist", "radioVideoId" ),
+    DeleteColumn( "Artist", "radioPlaylistId" ),
+)
+class From20To21Migration : AutoMigrationSpec
