+++ composeApp/src/androidMain/kotlin/me/knighthat/database/migrator/From21To22Migration.kt
@@ -1,0 +1,9 @@
+package me.knighthat.database.migrator
+
+import androidx.room.DeleteColumn
+import androidx.room.migration.AutoMigrationSpec
+
+@DeleteColumn.Entries(
+    DeleteColumn( "Artist", "info" )
+)
+class From21To22Migration : AutoMigrationSpec
