+++ composeApp/src/androidMain/kotlin/me/knighthat/database/migrator/From3To4Migration.kt
@@ -1,0 +1,9 @@
+package me.knighthat.database.migrator
+
+import androidx.room.DeleteTable
+import androidx.room.migration.AutoMigrationSpec
+
+@DeleteTable.Entries(
+    DeleteTable( tableName = "QueuedMediaItem" )
+)
+class From3To4Migration : AutoMigrationSpec
