+++ composeApp/src/androidMain/kotlin/me/knighthat/database/migrator/From7To8Migration.kt
@@ -1,0 +1,9 @@
+package me.knighthat.database.migrator
+
+import androidx.room.RenameColumn
+import androidx.room.migration.AutoMigrationSpec
+
+@RenameColumn.Entries(
+    RenameColumn("Song", "albumInfoId", "albumId")
+)
+class From7To8Migration : AutoMigrationSpec
